# Optimize peer memory halo exchange kernel for latency