This PR optimizes the peer memory halo exchange to minimize latency:
- Each CUDA thread performs its communication independently, which allows us to avoid cooperative group syncs. With enough threads, we avoid memory fences entirely. The downside is that we can only achieve 50% bandwidth since we pack data and flags into the same communication packets.
- Use an alternating double-buffer scheme, which allows push-only communication.
- The kernel does multiple rounds of communication in the cases where the number of SMs or the peer buffers are too small. This is important since the optimal peer memory requirements have increased by 4x, or more if the tensors are not nicely NHWC.  The only strict requirement is that the peer buffer sizes are aligned to 4096B.

Running on two DGX-1 V100s:
   | Version | Data type |   Kernel time |
   |-----------|-------|--------------------|
   | Baseline  | FP32  | 19.9 us |
   | Baseline  | FP16  |  17.2 us |
   | Optimized | FP32  | 15.2 us |
   | Optimized | FP16  | 11.4 us |

<details> <summary> Details </summary>

- I performed the halo exchange required for 3x3 convolution on a [1,336,336,64] NHWC tensor.
- To isolate the kernel time from the communication time, I delayed GPU 0 so it would always launch the kernel well after GPU 1.

</details>

Running the Mask R-CNN benchmark:

   | GPUs | Step time |
   |------|---------------|
   |    1 |       37.7 ms |
   |    2 |        34.8 ms |
   |    4 |        33.6 ms |

~The unit tests for the [halo exchanger](https://github.com/NVIDIA/apex/blob/master/apex/contrib/test/peer_memory/test_peer_halo_exchange_module.py) and the [bottleneck module](https://github.com/NVIDIA/apex/blob/master/apex/contrib/test/peer_memory/test_peer_halo_exchange_module.py) all pass for me on DGX-1 V100s.~ Pinging @thorjohnsen.