@timmoon10 2022-11-22
I've modified the bottleneck layer test to allocate more peer memory (since our peer memory requirement has grown 4x), and now it all passes for me.

I've also run some experiments on a DGX A100. Measuring the halo exchange kernel on two A100s:

   | Version | Data type |          Kernel time |
   |-----------|-------|--------------------|
   | Baseline  | FP32  | 28.3 us |
   | Baseline  | FP16  |   24.3 us |
   | Optimized | FP32  | 13.7 us |
   | Optimized | FP16  |  12.8 us |

Running the Mask R-CNN benchmark:

   | GPUs | Step time |
   |------|-----------|
   |    1 | 14.0 ms   |
   |    2 | 15.0 ms   |
   |    4 | 14.8 ms   |
   |    8 | 16.4 ms   |

It seems that with a sufficiently beefy GPU, we don't quite get strong scaling yet.